Creative portion, README.md and bug fixes.

+++ index.php
@@ -8,7 +8,19 @@
 <div class="container">
     <h1>File Sharing Site</h1>
 
-    <b>List User Files</b>
+    <h2>List User Files</h2>
+    <form class="center" method="GET">
+        <input class="button type <?php if (!isset($_GET["type"]) or $_GET["type"] == "All") echo "checked" ?>"
+               type="submit" name="type" value="All">
+        <input class="button type <?php if ($_GET["type"] == "Doc") echo "checked" ?>" type="submit" name="type"
+               value="Doc">
+        <input class="button type <?php if ($_GET["type"] == "PDF") echo "checked" ?>" type="submit" name="type"
+               value="PDF">
+        <input class="button type <?php if ($_GET["type"] == "Image") echo "checked" ?>" type="submit" name="type"
+               value="Image">
+        <input class="button type <?php if ($_GET["type"] == "Zip") echo "checked" ?>" type="submit" name="type"
+               value="Zip">
+    </form>
     <?php
     session_start();
     if (!isset($_SESSION["username"])) {
@@ -21,17 +33,30 @@
         mkdir($user_dir, 0777, true);
     }
     $files = scandir($user_dir);
+    if (isset($_GET["type"])) {
+        $type = $_GET["type"];
+    }
+
+    # Mapping of the types
+    $types["Doc"] = array('text/plain', 'application/msword', 'application/vnd.ms-excel', 'application/vnd.ms-powerpoint');
+    $types["PDF"] = array('application/pdf');
+    $types["Image"] = array('image/png', 'image/jpeg', 'image/gif');
+    $types["Zip"] = array('application/zip');
+
     foreach ($files as $file) {
         if (is_file($user_dir . $file)) {
-            echo "<p>" . htmlentities($file) . " ";
-            echo "<button onclick=\"window.location.href='openFile.php?file=" . htmlentities($file) . "'\">Open File</button>" . " ";
-            echo "<button onclick=\"window.location.href='deleteFile.php?file=" . htmlentities($file) . "'\">Delete File</button>";
-            echo "</p>";
+            # if the file type matches what we wanted
+            if (!isset($_GET["type"]) or $type == "All" or in_array(mime_content_type($user_dir . $file), $types[$type])) {
+                echo "<p>" . htmlentities($file) . " ";
+                echo "<button onclick=\"window.location.href='openFile.php?file=" . htmlentities($file) . "'\">Open File</button>" . " ";
+                echo "<button onclick=\"window.location.href='deleteFile.php?file=" . htmlentities($file) . "'\">Delete File</button>";
+                echo "</p>";
+            }
         }
     }
     ?>
 
-    <b>Upload File</b>
+    <h2>Upload File</h2>
     <form action="uploadFile.php" method="post" enctype="multipart/form-data">
         Select File to upload:
         <input class="button" type="file" name="fileToUpload" id="fileToUpload">
